refactores integrations:feature-gates logic into util function
