+++ src/sentry/static/sentry/app/views/organizationIntegrations/sentryAppDetailedView.tsx
@@ -14,13 +14,11 @@ import {
   uninstallSentryApp,
 } from 'app/actionCreators/sentryAppInstallations';
 import AsyncComponent from 'app/components/asyncComponent';
-import HookStore from 'app/stores/hookStore';
 import marked, {singleLineRenderer} from 'app/utils/marked';
 import InlineSvg from 'app/components/inlineSvg';
 import Tag from 'app/views/settings/components/tag';
 import {toPermissions} from 'app/utils/consolidatedScopes';
 import CircleIndicator from 'app/components/circleIndicator';
-import {Hooks} from 'app/types/hooks';
 import {
   IntegrationFeature,
   SentryApp,
@@ -28,6 +26,7 @@ import {
   SentryAppInstallation,
 } from 'app/types';
 import withOrganization from 'app/utils/withOrganization';
+import {getIntegrationFeatureGate} from 'app/utils/integrationUtil';
 import {UninstallButton} from '../settings/organizationDeveloperSettings/sentryApplicationRow/installButtons';
 
 type State = {
@@ -38,23 +37,6 @@ type State = {
 type Props = {
   organization: Organization;
 } & RouteComponentProps<{appSlug: string}, {}>;
-
-const defaultFeatureGateComponents = {
-  IntegrationFeatures: p =>
-    p.children({
-      disabled: false,
-      disabledReason: null,
-      ungatedFeatures: p.features,
-      gatedFeatureGroups: [],
-    }),
-  FeatureList: p => (
-    <ul>
-      {p.features.map((f, i) => (
-        <li key={i}>{f.description}</li>
-      ))}
-    </ul>
-  ),
-} as ReturnType<Hooks['integrations:feature-gates']>;
 
 class SentryAppDetailedView extends AsyncComponent<
   Props & AsyncComponent['props'],
@@ -189,9 +171,7 @@ class SentryAppDetailedView extends AsyncComponent<
       ),
     }));
 
-    const defaultHook = () => defaultFeatureGateComponents;
-    const featureHook = HookStore.get('integrations:feature-gates')[0] || defaultHook;
-    const {FeatureList, IntegrationFeatures} = featureHook();
+    const {FeatureList, IntegrationFeatures} = getIntegrationFeatureGate();
 
     const overview = sentryApp.overview || '';
     const featureProps = {organization, features};
